+++ src/main/java/community/whatever/onembackendjava/url/UrlShortenService.java
@@ -1,6 +1,28 @@
 package community.whatever.onembackendjava.url;
 
-public interface UrlShortenService {
-    String urlSearch(String key) ;
-    String keyCreate(String url) ;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UrlShortenService {
+
+    private  UrlShortenRepository urlShortenRepository ;
+
+    public UrlShortenService(UrlShortenRepository urlShortenRepository ) {
+        this.urlShortenRepository = urlShortenRepository;
+    }
+
+
+    public String urlSearch(String key) {
+        if(urlShortenRepository.existKey(key)){
+            return  urlShortenRepository.searchUrl(key) ;
+        }else{
+            throw new IllegalArgumentException("Invalid key");
+        }
+    }
+
+
+
+    public String keyCreate(String url) {
+        return urlShortenRepository.keyCreate(url);
+    }
 }
